GCP/02_Code/01_CloudFunction/requirements.txt
google-cloud-storage
google-cloud-firestore
google-cloud-speech
soundfile
numpy